some small corrections
