+++ bedboss/bedboss.py
@@ -14,22 +14,10 @@ from .const import (
     BED_FOLDER_NAME,
     BIGBED_FOLDER_NAME,
 )
+from utils import extract_file_name, standardize_genome_name, download_file
 from .exceptions import OSMException, GenomeException
 
 _LOGGER = logging.getLogger("bedboss")
-
-
-def download_file(url: str, path: str) -> NoReturn:
-    """
-    Download file from the url to specific location
-    :param url: URL of the file
-    :param path: Local path with filename
-    :return: NoReturn
-    """
-    _LOGGER.info(f"Downloading file: {url}")
-    _LOGGER.info(f"Downloading location: {path}")
-    urllib.request.urlretrieve(url, path)
-    _LOGGER.info(f"File has been downloaded successfully!")
 
 
 def get_osm_path(genome: str) -> Union[str, None]:
@@ -56,39 +44,6 @@ def get_osm_path(genome: str) -> Union[str, None]:
             os.makedirs(OPEN_SIGNAL_FOLDER)
         download_file(url=f"{OPEN_SIGNAL_URL}{osm_name}", path=osm_path)
     return osm_path
-
-
-def standard_genome_name(input_genome: str) -> str:
-    """
-    Standardizing user provided genome
-    :param input_genome: standardize user provided genome, so bedboss know what genome
-    we should use
-    :return: genome name string
-    """
-    _LOGGER.info(f"processing genome name...")
-    input_genome = input_genome.strip()
-    # TODO: we have to add more genome options and preprocessing of the string
-    if input_genome == "hg38" or input_genome == "GRCh38":
-        return "hg38"
-    elif input_genome == "hg19" or input_genome == "GRCh37":
-        return "hg19"
-    elif input_genome == "mm10":
-        return "mm10"
-    # else:
-    #     raise GenomeException("Incorrect genome assembly was provided")
-    else:
-        return input()
-
-
-def extract_file_name(file_path: str) -> str:
-    """
-    Extraction file name from file path
-    :param file_path: full file path
-    :return: file name without extension
-    """
-    file_name = os.path.basename(file_path)
-    file_name = file_name.split(".")[0]
-    return file_name
 
 
 def run_bedboss(
@@ -133,10 +88,7 @@ def run_bedboss(
     """
 
     file_name = extract_file_name(input_file)
-    genome = standard_genome_name(genome)
-    cwd = os.getcwd()
-    if not rfg_config:
-        rfg_config = os.path.join(cwd, "genome_config.yaml")
+    genome = standardize_genome_name(genome)
 
     # find/download open signal matrix
     if not open_signal_matrix:
@@ -150,6 +102,7 @@ def run_bedboss(
 
     output_bed = os.path.join(output_folder, BED_FOLDER_NAME, f"{file_name}.bed.gz")
     output_bigbed = os.path.join(output_folder, BIGBED_FOLDER_NAME)
+
     _LOGGER.info(f"output_bed = {output_bed}")
     _LOGGER.info(f"output_bigbed = {output_bigbed}")
